config: remove duplicated metadata cache initialization
Now the md5-dict format is selected if it's enabled in the
'cache-formats' setting in /path/to/repo/metadata/layout.conf, the pms
format is deprecated and only selected on directory discovery, otherwise
the final fallback is the local cache in
/var/cache/edb/cache/path/to/repo/*.

This means most repos without any cache-formats setting or old pms cache
dir will default to using the local cache. We might change this once we
change the 'cache-formats' setting to default to 'md5-dict' and/or drop
support for the 'pms' format.

Note that we currently only support using one type at once. Another idea
is to key off a features setting to use the local cache then fallback to
the secondary cache in the repo. This would be handy for remote repos
such as nfs-mount cases.